dashboard/summary/requirements.txt
streamlit==1.28.1
psycopg2-binary==2.9.9
pandas==2.1.3
python-dotenv==1.0.0